+++ SharedLibrary/Enums/LogEventType.cs
@@ -5,4 +5,5 @@ public enum LogEventType
     Bounty,
     Combat,
     Jump,
+    Other,
 }