+++ UI.avalonia/Views/LogView.axaml.cs
@@ -1,4 +1,6 @@
 using Avalonia.Controls;
+using Microsoft.Extensions.DependencyInjection;
+using UI.avalonia.ViewModels;
 
 namespace UI.avalonia.Views;
 
@@ -7,6 +9,8 @@ public partial class LogView : UserControl
     public LogView()
     {
         InitializeComponent();
-        // TODO: Port full LogView logic from WPF project
+
+        if (!Design.IsDesignMode && App.ServiceProvider is { } serviceProvider)
+            DataContext = serviceProvider.GetRequiredService<LogViewModel>();
     }
 }